Initial commit
This is pretty much verbatim the Android Notepad Version 3 tutorial with
some buttons added. to cancel a change and delete a message. The
messages are time stamped and the title is removed (as it is more
imperative to see what the message may have in it rather than force a
user to type in a title that has nothing to do with the message).