Changed menu handling technique, added a lot more comments
Reading the Android dev site led me to use the menu.xml format instead
of hand creating the menus. So I rewrote all the menus, even added a
menu for the button commands in the edit screen. I also commented all
the code, since I hadn't done it yet. Refactored the code a little as
well.