Added dictation
Added dictation button to handle speech input. (yes, this is handled by
the device through the use of the record button in the keyboard but it
seemed neat to play with).

Note that there seems to be an Android issue if there is no returned
(decipherable) string for the audio. I tried but could not find a place
to trap the exception (it occurs in their main for the audio capture
handler). I'm not sure if the problem is my device or Android.